Multi port support fixes  - fix useless aws network interface creation  - improve windows script for multiple port support  - fix hybrid use of level VM and tenant  - delete provider network interface in case of multi ports

+++ doc/source/scripts/connect.ps1
@@ -2,13 +2,8 @@ Stop-Process -processname "openvpn-gui"
 net stop OpenVpnService
 net stop OpenVPNServiceInteractive
 
-netsh interface set interface "Ethernet"   disable
-netsh interface set interface "Ethernet 2" disable
-netsh interface set interface "Ethernet 3" disable
-netsh interface set interface "Ethernet 4" disable
-netsh interface set interface "Ethernet 5" disable
-netsh interface ipv4 set address name=Ethernet source=dhcp
 netsh interface set interface "Ethernet" enable
+netsh interface ipv4 set address name=Ethernet source=dhcp
 
 Start-Sleep -s 5
 
@@ -19,8 +14,10 @@ $userdata = ConvertFrom-StringData -StringData $UserData
 # format:
 # hsservers0 = xxx.xxx.xxx.xxx, xxx.xxx.xxx.xxx
 # mac0 = 00:00:00:00:00:00
+# port0 = XXXX
 # hsservers1 = xxx.xxx.xxx.xxx, xxx.xxx.xxx.xxx
 # mac1 = 00:00:00:00:00:00
+# port1 = XXXX
 
 $i = 0
 $hsservers = $userdata."hsservers$i"
@@ -28,31 +25,32 @@ $mac = $userdata."mac$i"
 $port = $userdata."port$i"
 $need_restart = $false
 
+$n_eth = "Ethernet"
+echo "n_eth: '$n_eth'"
+
+$tap_bin_dir = "C:\Program Files\Tap-Windows\bin"
+$openvpn_conf_dir = "C:\Program Files\OpenVPN\config"
+
 While ("$mac" -ne "") {
 
     echo "mac: '$mac'"
     echo "hsservers: '$hsservers'"
+    echo "port: '$port'"
+
     $hsservers = $hsservers -split ","
-    $openvpn_bin_dir = "C:\Program Files\OpenVPN\bin"
-    $openvpn_conf_dir = "C:\Program Files\OpenVPN\config"
     $openvpn_file_conf = "c-hs-" + $i + ".ovpn"
     
-    if ($i -eq 0) {
-        $n_eth = "Ethernet"
-        $vpn_eth = "Ethernet 2"
-    } else {
-        $n_ind = 2 * $i + 1
-        $n_eth = "Ethernet " + $n_ind
-        $vpn_ind = 2 * $i + 2
-        $vpn_eth = "Ethernet " + $vpn_ind
-    }
-    echo "n_eth: '$n_eth'"
+    $vpn_ind = $i + 2
+    $vpn_eth = "Ethernet " + $vpn_ind
+
     echo "vpn_eth: '$vpn_eth'"
-    netsh interface ipv4 set address name="$n_eth" source=dhcp
-    netsh interface set interface "$n_eth" enable
+    $vpn_nic_exist = openvpn --show-adapters | findstr /R /C:"$vpn_eth"
+    if ("$vpn_nic_exist" -eq "") {
+        cd $tap_bin_dir
+        .\tapinstall.exe install "C:\Program Files\TAP-Windows\driver\OemVista.inf" tap0901
+    }
 
     cd $openvpn_conf_dir
-
     "client" | Out-File -FilePath $openvpn_file_conf -enc UTF8
     "dev tap" | Out-File -FilePath $openvpn_file_conf -Append -enc UTF8
     "dev-node ""$vpn_eth""" | Out-File -FilePath $openvpn_file_conf -Append -enc UTF8
@@ -72,7 +70,7 @@ While ("$mac" -ne "") {
     "key client.key" | Out-File -FilePath $openvpn_file_conf -Append -enc UTF8
     "verb 3" | Out-File -FilePath $openvpn_file_conf -Append -enc UTF8
     
-    # serach the reg key
+    # search the reg key
     $d_id_0 = "ROOT\NET\000" + $i
     echo "d_id_0: '$d_id_0'"
     for ($a = 10; $a -le 30; $a++){
@@ -83,44 +81,41 @@ While ("$mac" -ne "") {
             Break
         }
     }
-    $cur_mac = (Get-ItemProperty -Path "$reg_key").MAC
-    echo "mac: '$mac'"
-    echo "cur_mac: '$cur_mac'"
-    echo "reg_key: '$reg_key'"
-    If ($cur_mac -ne $mac){
-       netsh interface set interface "$vpn_eth" disable
-       echo "Not same mac"
-       Remove-ItemProperty -Path $reg_key -Name "MAC"
-       New-ItemProperty -Path $reg_key -Name "MAC" -Value "$mac"
-       netsh interface set interface "$vpn_eth" enable
-       $need_restart = $true
+    if ($d_id -eq $d_id_0) {
+        $cur_mac = (Get-ItemProperty -Path "$reg_key").MAC
+        echo "mac: '$mac'"
+        echo "cur_mac: '$cur_mac'"
+        echo "reg_key: '$reg_key'"
+        If ($cur_mac -ne $mac){
+           netsh interface set interface "$vpn_eth" disable
+           echo "Not same mac"
+           Remove-ItemProperty -Path $reg_key -Name "MAC"
+           New-ItemProperty -Path $reg_key -Name "MAC" -Value "$mac"
+           $need_restart = $true
+        }
     }
 
-    $ip = netsh int ip show config name="$n_eth" | findstr /R /C:"IP Address"
-    $ip = $ip.split(":")[1].trim()
-    $netmask = netsh int ip show config name="$n_eth" | findstr /R /C:"Subnet"
-    $netmask = $netmask.split(":")[1].trim().split(" ")[2].trim().split(")")[0].trim()
-    echo "netmask: '$netmask'"
-    netsh interface ipv4 set address name="$n_eth" static $ip $netmask
-    netsh interface set interface "$n_eth" enable
     netsh interface set interface "$vpn_eth" enable
     
-#    cd $openvpn_bin_dir
-#    openvpn-gui.exe --connect c-hs.ovpn --config_dir $openvpn_conf_dir
-
     $i = $i + 1
     $hsservers = $userdata."hsservers$i"
     $mac = $userdata."mac$i"
     $port = $userdata."port$i"
 }
 
+# disable the dhcp and default route
+$ip = netsh int ip show config name="$n_eth" | findstr /R /C:"IP Address"
+$ip = $ip.split(":")[1].trim()
+$netmask = netsh int ip show config name="$n_eth" | findstr /R /C:"Subnet"
+$netmask = $netmask.split(":")[1].trim().split(" ")[2].trim().split(")")[0].trim()
+echo "netmask: '$netmask'"
+netsh interface ipv4 set address name="$n_eth" static $ip $netmask
+netsh interface set interface "$n_eth" enable
+
 if ($need_restart) {
     Restart-Computer
 }
 
-net start OpenVpnService
-
-Start-Sleep -s 5
-
+# restart openvpn service
 net stop OpenVpnService
 net start OpenVpnService
